MDEV-31404 don't use ulong for sysvars
it's not portable

+++ sql/mysqld.cc
@@ -483,7 +483,7 @@ ulonglong slave_type_conversions_options;
 ulong thread_cache_size=0;
 ulonglong binlog_cache_size=0;
 ulonglong binlog_file_cache_size=0;
-ulong slave_connections_needed_for_purge;
+uint slave_connections_needed_for_purge;
 ulonglong max_binlog_cache_size=0;
 ulong slave_max_allowed_packet= 0;
 double slave_max_statement_time_double;
